feat: change component name and clean the code of test and img component
